moved python script from comment to its own file. also fixed a compile error
