+++ Apps/Inc/VoltageToTemp.h
@@ -1,63 +1,14 @@
 /* Copyright (c) 2021 UT Longhorn Racing Solar */
 
+#ifndef __VOLTAGETOTEMP_H__
+#define __VOLTAGETOTEMP_H__
+
 #include <stdint.h>
 
 // Lookup table from TI's datasheeet can be downloaded here:
 // https://www.ti.com/lit/zip/snic009
 
-// Values generated by:
-/*
-import pandas as pd
-
-UNKNOWN = 10**9
-
-def interpolate(idx, predecessor, successor):
-    delta = idx - predecessor[0]
-    slope = (successor[1] - predecessor[1]) / (successor[0] - predecessor[0])
-    return int(delta * slope + predecessor[1])
-
-def main():
-    # read in TI's lookup table
-    df = pd.read_csv("LMT87_LookUpTable.csv")
-    partial_table = {}
-    for row in range(len(df)):
-        partial_table[int(df["millivolts"][row])] = int(df["temperature"][row])
-    # populate our table with the values TI gives us
-    table = [UNKNOWN] * 3301
-    for val in range(0, 3301):
-        if val in partial_table.keys():
-            table[val] = partial_table[val] * 1000
-    # interpolate the rest of the values
-    predecessor = (UNKNOWN, UNKNOWN)
-    successor = (UNKNOWN, UNKNOWN)
-    vals_to_calc = []
-    for idx, val in enumerate(table):
-        if val == UNKNOWN:
-            vals_to_calc.append(idx)
-        else:
-            if successor[0] != UNKNOWN:
-                predecessor = successor
-            successor = (idx, val)
-            if predecessor[0] != UNKNOWN:
-                while len(vals_to_calc) > 0:
-                    update_idx = vals_to_calc.pop()
-                    print("interpolating mV = " + str(update_idx))
-                    table[update_idx] = interpolate(update_idx, predecessor, successor)
-    # interpolate ending values
-    while len(vals_to_calc) > 0:
-                    update_idx = vals_to_calc.pop()
-                    print("interpolating mV = " + str(update_idx))
-                    table[update_idx] = interpolate(update_idx, predecessor, successor)
-    # write the lookup table to a file as a C array
-    with open("VoltageToTemp.c", "w") as file:
-        file.write("const int32_t voltToTemp[] = {\n")
-        for val in table:
-            file.write("    " + str(int(val)) + ",\n")
-        file.write("};\n")
-
-if __name__ == "__main__":
-    main()
-*/
+// Values generated by voltToTemp.py
 
 // lookup table for converting ADC voltages to temperatures
 const int32_t voltToTemp[] = {
@@ -3363,3 +3314,5 @@ const int32_t voltToTemp[] = {
     -52000,
     -52090,
 };
+
+#endif
